Add the languagepack list functionality to the dashboard
The dashboard plugin currently does not support languagepack
functionalities. The patch adds support to the plugin for listing
languagepacks.

To test the patch, download the patch, and replace the changed files in
the /opt/stack/solum-dashboard directory in the devstack setup.

Also, here is a link to a screenchot of what the page looks like:
http://pasteboard.co/HYOOPBgO.png

Closes-Bug: #1597373
Change-Id: I54c9adeb1c27dd6b770704aa3ddf715e2069a9f8