Add the languagepack delete functionality to the dashboard
The dashboard plugin currently does not support the option to delete a
selected languagepack. This patch adds this option to the dashboard
plugin.

Also, here is a link to a screenchot of what the page looks like before
confirming the delete: http://pasteboard.co/HU1vtWlb.png, and after
confirming the delete: http://pasteboard.co/HT6AoO5g.png

Change-Id: I6d6a1c940c35262b2b2531026a3b82a74e47b69e
Closes-Bug: #1600616