Add the languagepack create functionality to the dashboard
The dashboard plugin currently does not support the option to create a
new languagepack. This patch adds this option to the dashboard plugin.

Also, here is a link to a screenchot of what the form for creation looks
like: http://pasteboard.co/4hlr2GO3z.png, and what the page after
creation of the languagepack looks like:
http://pasteboard.co/I0HuPcLh.png

Change-Id: I3c53ae0cdfcbd4bb8f99d9003ccfebc5b3e5c7ee
Closes-Bug: #1602874