Fix lower-constraint deps handling
When the lower-constraints tox target was added, it was assumed the
install_command was just running the install and that the dependencies
and constraints were being set using "deps = ".

This fixed the install_command and deps to follow the expected pattern
so the lower-constraints job actual does install the lower constraints.

Change-Id: Ibb93f81ed23833f64730ca851a88114d4f0e0e63
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>