Drop django-babel from requirement.txt
We have switched the babel extractor for Django from django-babel
to enmerkar recently. solum-dashboard requirements.txt needs to
sync the change. enmerkar, a successor of django-babel will be
installed via horizon, so we don't need to have it in requirements.txt
explicitly.

lower-constrants.txt is updated to use enmerkar accordingly.

Change-Id: Idfc3c4e39f616723e744cf13037d72bebc1026db